+++ NEWS.md
@@ -1,6 +1,7 @@
 # chromConverter 0.1.1
 
-* Added `read_fid` function for parsing Shimadzu `.txt` files.
+* Added `read_fid` function for parsing Shimadzu GC-FID files.
+* Added `read_chromeleon` function for parsing Chromeleon HPLC-UV files.
 
 # chromConverter 0.1.0